some dependency fixes for tests
