+++ .github/workflows/test_airflow.yml
@@ -41,7 +41,7 @@ jobs:
           key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}-airflow-runner
 
       - name: Install dependencies
-        run: poetry install --no-interaction --with airflow -E duckdb -E bigquery
+        run: poetry install --no-interaction --with airflow -E duckdb
 
       - run: |
           poetry run pytest tests/helpers/airflow_tests
